# examples/openapi-demo/openapi-rest.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API (REST Mode)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/createUser": {
      "post": {
        "summary": "POST createUser",
        "operationId": "createUser",
        "tags": [
          "RPC"
        ],
        "requestBody": {
          "description": "Request body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/getUser": {
      "get": {
        "summary": "GET getUser",
        "operationId": "getUser",
        "tags": [
          "RPC"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Request parameter",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}